+++ queue.h
@@ -7,10 +7,12 @@ extern "C"
 {
 #endif
 
+#include "inttypes.h"
+
 typedef struct
 {
     struct nav_cell *cell;
-    int n;
+    int8_t n;
 } nav_queue_cell;
 
 typedef struct
